use properties for static classes + reorganize lua.type package
