+++ src/main/java/me/hugeblank/allium/lua/api/ConfigLib.java
@@ -4,8 +4,8 @@ import com.google.common.base.Strings;
 import com.google.gson.JsonElement;
 import me.hugeblank.allium.loader.Script;
 import me.hugeblank.allium.loader.ScriptResource;
-import me.hugeblank.allium.lua.type.LuaWrapped;
-import me.hugeblank.allium.lua.type.OptionalArg;
+import me.hugeblank.allium.lua.type.annotation.LuaWrapped;
+import me.hugeblank.allium.lua.type.annotation.OptionalArg;
 import me.hugeblank.allium.util.FileHelper;
 import org.squiddev.cobalt.Constants;
 import org.squiddev.cobalt.LuaError;
