+++ src/main/java/me/hugeblank/allium/lua/type/annotation/CoerceToBound.java
@@ -1,9 +1,9 @@
-package me.hugeblank.allium.lua.type;
+package me.hugeblank.allium.lua.type.annotation;
 
 import java.lang.annotation.*;
 
 @Target(ElementType.TYPE_USE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CoerceToNative {
+public @interface CoerceToBound {
 }
