+++ src/main/java/me/hugeblank/allium/lua/type/annotation/CoerceToNative.java
@@ -1,9 +1,9 @@
-package me.hugeblank.allium.lua.type;
+package me.hugeblank.allium.lua.type.annotation;
 
 import java.lang.annotation.*;
 
 @Target(ElementType.TYPE_USE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CoerceToBound {
+public @interface CoerceToNative {
 }
